feat: replace current main.go for refactored one
