--- cmd/auto-update/main2.go
@@ -1,111 +1,0 @@
-package main
-
-import (
-	"context"
-	"os"
-	"time"
-
-	"github.com/canopy-network/canopy/cmd/cli"
-	"github.com/canopy-network/canopy/cmd/rpc"
-	"github.com/canopy-network/canopy/lib"
-)
-
-var (
-	defaultCheckPeriod = time.Minute * 30
-)
-
-const (
-	snapshotFileName    = "snapshot.tar.gz"
-	snapshotMetadataKey = "snapshot"
-	httpClientTimeout   = time.Second * 10
-
-	// program defaults
-	defaultRepoName  = "canopy"
-	defaultRepoOwner = "rem1niscence"
-	defaultBinPath   = "./cli"
-
-	githubAPIBaseURL = "https://api.github.com"
-)
-
-func main() {
-	// check if start was called
-	if len(os.Args) < 2 || os.Args[1] != "start" {
-		log.Fatalf("invalid input %v only `start` command is allowed", os.Args)
-	}
-	// get configs and logger
-	configs, logger := getConfigs()
-	// setup the dependencies
-	updater := NewUpdateManager(configs.Updater, logger, rpc.SoftwareVersion)
-	snapshot := NewSnapshotManager(configs.Snapshot)
-	supervisor := NewProcessSupervisor(configs.Supervisor, logger)
-	coordinator := NewCoordinator(configs.Coordinator, updater, supervisor, snapshot, logger)
-	// start the program
-	if err := supervisor.Start(); err != nil {
-		log.Fatalf("failed to start supervisor: %v", err)
-	}
-	// start the update loop
-	if err := coordinator.StartUpdateLoop(context.Background()); err != nil {
-		log.Fatalf("failed to stop update loop: %v", err)
-	}
-}
-
-// Configs holds the configuration for the updater, snapshotter, and process supervisor.
-type Configs struct {
-	Updater     *UpdaterConfig
-	Snapshot    *SnapshotConfig
-	Supervisor  *SupervisorConfig
-	Coordinator *CoordinatorConfig
-	LoggerI     lib.LoggerI
-}
-
-func getConfigs() (*Configs, lib.LoggerI) {
-	config, _ := cli.InitializeDataDirectory(cli.DataDir, lib.NewDefaultLogger())
-	l := lib.NewLogger(lib.LoggerConfig{
-		Level:      config.GetLogLevel(),
-		Structured: config.Structured,
-		JSON:       config.JSON,
-	})
-
-	binPath := envOrDefault("CANOPY_BIN_PATH", defaultBinPath)
-
-	updater := &UpdaterConfig{
-		RepoName:  envOrDefault("CANOPY_REPO_NAME", defaultRepoName),
-		RepoOwner: envOrDefault("CANOPY_REPO_OWNER", defaultRepoOwner),
-		BinPath:   binPath,
-		CheckTime: defaultCheckPeriod,
-	}
-
-	snapshot := &SnapshotConfig{
-		canopy: config,
-		URLs:   snapshotURLs,
-		Name:   snapshotFileName,
-	}
-
-	supervisor := &SupervisorConfig{
-		canopy:  config,
-		BinPath: binPath,
-	}
-
-	coordinator := &CoordinatorConfig{
-		Canopy:      config,
-		CheckPeriod: defaultCheckPeriod,
-	}
-
-	return &Configs{
-		Updater:     updater,
-		Snapshot:    snapshot,
-		Supervisor:  supervisor,
-		Coordinator: coordinator,
-		LoggerI:     l,
-	}, l
-}
-
-// envOrDefault returns the value of the environment variable with the given key,
-// or the default value if the variable is not set.
-func envOrDefault(key, defaultValue string) string {
-	value := os.Getenv(key)
-	if value == "" {
-		return defaultValue
-	}
-	return value
-}
